integrate new version of Jimp

+++ routes/index.js
@@ -46,7 +46,7 @@ router.post('/score', function (req, response, next) {
 
     }
     if (req.body && req.body.generatePart) {
-        scoreSplitter.generatePart(req.body.pdfName, req.body.part, req.body.zonesStr,parseInt(req.body.margin), parseInt(req.body.imgScaleCoef),function (error, result) {
+        scoreSplitter.generatePart(req.body.pdfName, req.body.part, req.body.zonesStr,parseInt(req.body.margin), parseFloat(req.body.imgScaleCoef),function (error, result) {
             processResponse(response, error, result)
         })
